add check if user chose file (uploaded file) before remove the filr

+++ server/routes/registerRouter.js
@@ -27,7 +27,10 @@ router.post(
     async (req, res, next) => {
       const errors = validationResult(req);
       if (!errors.isEmpty()) {
-        await fs.unlink(req.file.path); // remove previously uploaded file
+        // if chosen file (uploaded file)
+        if (req.file) {
+          await fs.unlink(req.file.path); // remove previously uploaded file
+        }
         req.flash('validation_errors', errors.array());
         return res.redirect(req.originalUrl);
       }
